goalkeeper/requirements.txt
Django==2.2.1
djangorestframework==3.9.3
django-widget-tweaks==1.4.3
django-filter==2.1.0
faker==2.0.0
